+++ services/alert/handlers.go
@@ -481,6 +481,21 @@ func (h *matchHandler) Handle(event alert.Event) {
 	}
 }
 
+var changedFuncSignature = map[stateful.Domain]ast.ValueType{}
+var levelFuncSignature = map[stateful.Domain]ast.ValueType{}
+var nameFuncSignature = map[stateful.Domain]ast.ValueType{}
+var taskNameFuncSignature = map[stateful.Domain]ast.ValueType{}
+var durationFuncSignature = map[stateful.Domain]ast.ValueType{}
+
+func init() {
+	d := stateful.Domain{}
+	changedFuncSignature[d] = ast.TBool
+	levelFuncSignature[d] = ast.TInt
+	nameFuncSignature[d] = ast.TString
+	taskNameFuncSignature[d] = ast.TString
+	durationFuncSignature[d] = ast.TDuration
+}
+
 func (h *matchHandler) match(event alert.Event) (bool, error) {
 	// Populate scope
 	h.scope.Reset()
@@ -488,47 +503,62 @@ func (h *matchHandler) match(event alert.Event) (bool, error) {
 	h.logger.Printf("D! match %+v", h)
 
 	if h.usesChanged {
-		h.scope.SetDynamicMethod(changedFunc, func(self interface{}, args ...interface{}) (interface{}, error) {
-			if len(args) != 0 {
-				return nil, fmt.Errorf("%s takes no arguments", changedFunc)
-			}
-			return event.State.Level != event.PreviousState().Level, nil
+		h.scope.SetDynamicFunc(changedFunc, &stateful.DynamicFunc{
+			F: func(args ...interface{}) (interface{}, error) {
+				if len(args) != 0 {
+					return nil, fmt.Errorf("%s takes no arguments", changedFunc)
+				}
+				return event.State.Level != event.PreviousState().Level, nil
+			},
+			Sig: changedFuncSignature,
 		})
 	}
 
 	if h.usesLevel {
-		h.scope.SetDynamicMethod(levelFunc, func(self interface{}, args ...interface{}) (interface{}, error) {
-			if len(args) != 0 {
-				return nil, fmt.Errorf("%s takes no arguments", levelFunc)
-			}
-			return int64(event.State.Level), nil
+		h.scope.SetDynamicFunc(levelFunc, &stateful.DynamicFunc{
+			F: func(args ...interface{}) (interface{}, error) {
+				if len(args) != 0 {
+					return nil, fmt.Errorf("%s takes no arguments", levelFunc)
+				}
+				return int64(event.State.Level), nil
+			},
+			Sig: levelFuncSignature,
 		})
 	}
 
 	if h.usesName {
-		h.scope.SetDynamicMethod(nameFunc, func(self interface{}, args ...interface{}) (interface{}, error) {
-			if len(args) != 0 {
-				return nil, fmt.Errorf("%s takes no arguments", nameFunc)
-			}
-			return event.Data.Name, nil
+		h.scope.SetDynamicFunc(nameFunc, &stateful.DynamicFunc{
+			F: func(args ...interface{}) (interface{}, error) {
+				if len(args) != 0 {
+					return nil, fmt.Errorf("%s takes no arguments", nameFunc)
+				}
+				return event.Data.Name, nil
+			},
+			Sig: nameFuncSignature,
 		})
 	}
 
 	if h.usesTaskName {
-		h.scope.SetDynamicMethod(taskNameFunc, func(self interface{}, args ...interface{}) (interface{}, error) {
-			if len(args) != 0 {
-				return nil, fmt.Errorf("%s takes no arguments", taskNameFunc)
-			}
-			return event.Data.TaskName, nil
+		h.scope.SetDynamicFunc(taskNameFunc, &stateful.DynamicFunc{
+			F: func(args ...interface{}) (interface{}, error) {
+				if len(args) != 0 {
+					return nil, fmt.Errorf("%s takes no arguments", taskNameFunc)
+				}
+				return event.Data.TaskName, nil
+			},
+			Sig: taskNameFuncSignature,
 		})
 	}
 
 	if h.usesDuration {
-		h.scope.SetDynamicMethod(durationFunc, func(self interface{}, args ...interface{}) (interface{}, error) {
-			if len(args) != 0 {
-				return nil, fmt.Errorf("%s takes no arguments", durationFunc)
-			}
-			return event.State.Duration, nil
+		h.scope.SetDynamicFunc(durationFunc, &stateful.DynamicFunc{
+			F: func(args ...interface{}) (interface{}, error) {
+				if len(args) != 0 {
+					return nil, fmt.Errorf("%s takes no arguments", durationFunc)
+				}
+				return event.State.Duration, nil
+			},
+			Sig: durationFuncSignature,
 		})
 	}