+++ tick/stateful/eval_float_node.go
@@ -11,7 +11,7 @@ type EvalFloatNode struct {
 	Float64 float64
 }
 
-func (n *EvalFloatNode) Type(scope ReadOnlyScope, executionState ExecutionState) (ast.ValueType, error) {
+func (n *EvalFloatNode) Type(scope ReadOnlyScope) (ast.ValueType, error) {
 	return ast.TFloat, nil
 }
 
@@ -30,15 +30,23 @@ func (n *EvalFloatNode) EvalString(scope *Scope, executionState ExecutionState) 
 func (n *EvalFloatNode) EvalBool(scope *Scope, executionState ExecutionState) (bool, error) {
 	return false, ErrTypeGuardFailed{RequestedType: ast.TBool, ActualType: ast.TFloat}
 }
+
 func (n *EvalFloatNode) EvalRegex(scope *Scope, executionState ExecutionState) (*regexp.Regexp, error) {
 	return nil, ErrTypeGuardFailed{RequestedType: ast.TRegex, ActualType: ast.TFloat}
 }
+
 func (n *EvalFloatNode) EvalTime(scope *Scope, executionState ExecutionState) (time.Time, error) {
 	return time.Time{}, ErrTypeGuardFailed{RequestedType: ast.TTime, ActualType: ast.TFloat}
 }
+
 func (n *EvalFloatNode) EvalDuration(scope *Scope, executionState ExecutionState) (time.Duration, error) {
 	return 0, ErrTypeGuardFailed{RequestedType: ast.TDuration, ActualType: ast.TFloat}
 }
+
+func (n *EvalFloatNode) EvalMissing(scope *Scope, executionState ExecutionState) (*ast.Missing, error) {
+	return nil, ErrTypeGuardFailed{RequestedType: ast.TMissing, ActualType: ast.TFloat}
+}
+
 func (n *EvalFloatNode) IsDynamic() bool {
 	return false
 }