+++ tick/stateful/eval_function_node.go
@@ -3,6 +3,7 @@ package stateful
 import (
 	"fmt"
 	"regexp"
+	"strings"
 	"time"
 
 	"github.com/influxdata/kapacitor/tick/ast"
@@ -31,29 +32,29 @@ func NewEvalFunctionNode(funcNode *ast.FunctionNode) (*EvalFunctionNode, error) 
 	return evalFuncNode, nil
 }
 
-func (n *EvalFunctionNode) Type(scope ReadOnlyScope, executionState ExecutionState) (ast.ValueType, error) {
-	f := executionState.Funcs[n.funcName]
+func (n *EvalFunctionNode) Type(scope ReadOnlyScope) (ast.ValueType, error) {
+	f := lookupFunc(n.funcName, builtinFuncs, scope)
 	if f == nil {
 		return ast.InvalidType, fmt.Errorf("undefined function: %q", n.funcName)
 	}
 	signature := f.Signature()
 
 	domain := Domain{}
-	if len(n.argsEvaluators) > len(domain) {
-	}
-
 	for i, argEvaluator := range n.argsEvaluators {
-		t, err := argEvaluator.Type(scope, executionState)
+		t, err := argEvaluator.Type(scope)
 		if err != nil {
 			return ast.InvalidType, fmt.Errorf("Failed to handle %v argument: %v", i+1, err)
 		}
 		domain[i] = t
 	}
 
+	if gotLen, expLen := len(n.argsEvaluators), len(domain); gotLen > expLen {
+		return ast.InvalidType, ErrWrongFuncSignature{Name: n.funcName, DomainProvided: domain, Func: f}
+	}
+
 	retType, ok := signature[domain]
 	if !ok {
-		// TODO: Cleanup error returned here
-		return ast.InvalidType, fmt.Errorf("Wrong function signature")
+		return ast.InvalidType, ErrWrongFuncSignature{Name: n.funcName, DomainProvided: domain, Func: f}
 	}
 
 	return retType, nil
@@ -75,17 +76,9 @@ func (n *EvalFunctionNode) callFunction(scope *Scope, executionState ExecutionSt
 		args = append(args, value)
 	}
 
-	f := executionState.Funcs[n.funcName]
-
-	// Look for function on scope
+	f := lookupFunc(n.funcName, executionState.Funcs, scope)
 	if f == nil {
-		if dm := scope.DynamicMethod(n.funcName); dm != nil {
-			f = dynamicMethodFunc{dm: dm}
-		}
-	}
-
-	if f == nil {
-		return nil, fmt.Errorf("undefined function: %q", n.funcName)
+		return ast.InvalidType, fmt.Errorf("undefined function: %q", n.funcName)
 	}
 
 	ret, err := f.Call(args...)
@@ -187,10 +180,22 @@ func (n *EvalFunctionNode) EvalBool(scope *Scope, executionState ExecutionState)
 	return false, ErrTypeGuardFailed{RequestedType: ast.TBool, ActualType: ast.TypeOf(refValue)}
 }
 
+func (n *EvalFunctionNode) EvalMissing(scope *Scope, executionState ExecutionState) (*ast.Missing, error) {
+	refValue, err := n.callFunction(scope, executionState)
+	if err != nil {
+		return nil, err
+	}
+
+	if missingValue, isMissing := refValue.(*ast.Missing); isMissing {
+		return missingValue, nil
+	}
+	return nil, ErrTypeGuardFailed{RequestedType: ast.TMissing, ActualType: ast.TypeOf(refValue)}
+}
+
 // eval - generic evaluation until we have reflection/introspection capabillities so we can know the type of args
 // and return type, we can remove this entirely
 func eval(n NodeEvaluator, scope *Scope, executionState ExecutionState) (interface{}, error) {
-	retType, err := n.Type(scope, executionState)
+	retType, err := n.Type(scope)
 	if err != nil {
 		return nil, err
 	}
@@ -210,18 +215,32 @@ func eval(n NodeEvaluator, scope *Scope, executionState ExecutionState) (interfa
 		return n.EvalTime(scope, executionState)
 	case ast.TDuration:
 		return n.EvalDuration(scope, executionState)
+	case ast.TMissing:
+		v, err := n.EvalMissing(scope, executionState)
+		if err != nil && !strings.Contains(err.Error(), "missing value") {
+			return v, err
+		}
+		return v, nil
 	default:
 		return nil, fmt.Errorf("function arg expression returned unexpected type %s", retType)
 	}
 
 }
 
-type dynamicMethodFunc struct {
-	dm DynamicMethod
-}
+func lookupFunc(name string, funcs Funcs, scope ReadOnlyScope) Func {
+	f := funcs[name]
+	if f != nil {
+		return f
+	}
 
-func (dmf dynamicMethodFunc) Call(args ...interface{}) (interface{}, error) {
-	return dmf.dm(nil, args...)
-}
-func (dmf dynamicMethodFunc) Reset() {
+	df := scope.DynamicFunc(name)
+	if df != nil {
+		return df
+	}
+
+	// Return nil here explicitly so its a nil Func not nil *DynamicFunc
+	// returning scope.DynamicFunc(name) caused nil check in callFunction
+	// f == nil to evaluate to false even though f was nil. This was weird
+	// enough that I felt it warranted a comment.
+	return nil
 }