+++ tick/stateful/eval_reference_node.go
@@ -13,7 +13,7 @@ type EvalReferenceNode struct {
 }
 
 // getReferenceValue - core method for evaluating function where all NodeEvaluator methods should use
-func (n *EvalReferenceNode) getReferenceValue(scope *Scope, executionState ExecutionState) (interface{}, error) {
+func (n *EvalReferenceNode) getReferenceValue(scope *Scope) (interface{}, error) {
 	value, err := scope.Get(n.Node.Reference)
 	if err != nil {
 		return nil, err
@@ -25,8 +25,8 @@ func (n *EvalReferenceNode) getReferenceValue(scope *Scope, executionState Execu
 	return value, nil
 }
 
-func (n *EvalReferenceNode) Type(scope ReadOnlyScope, executionState ExecutionState) (ast.ValueType, error) {
-	value, err := n.getReferenceValue(scope.(*Scope), executionState)
+func (n *EvalReferenceNode) Type(scope ReadOnlyScope) (ast.ValueType, error) {
+	value, err := n.getReferenceValue(scope.(*Scope))
 	if err != nil {
 		return ast.InvalidType, err
 	}
@@ -39,7 +39,7 @@ func (n *EvalReferenceNode) IsDynamic() bool {
 }
 
 func (n *EvalReferenceNode) EvalRegex(scope *Scope, executionState ExecutionState) (*regexp.Regexp, error) {
-	refValue, err := n.getReferenceValue(scope, executionState)
+	refValue, err := n.getReferenceValue(scope)
 	if err != nil {
 		return nil, err
 	}
@@ -48,11 +48,16 @@ func (n *EvalReferenceNode) EvalRegex(scope *Scope, executionState ExecutionStat
 		return regexValue, nil
 	}
 
+	refType := ast.TypeOf(refValue)
+	if refType == ast.TMissing {
+		return nil, fmt.Errorf("reference \"%s\" is missing value", n.Node.Reference)
+	}
+
 	return nil, ErrTypeGuardFailed{RequestedType: ast.TRegex, ActualType: ast.TypeOf(refValue)}
 }
 
 func (n *EvalReferenceNode) EvalTime(scope *Scope, executionState ExecutionState) (time.Time, error) {
-	refValue, err := n.getReferenceValue(scope, executionState)
+	refValue, err := n.getReferenceValue(scope)
 	if err != nil {
 		return time.Time{}, err
 	}
@@ -61,11 +66,16 @@ func (n *EvalReferenceNode) EvalTime(scope *Scope, executionState ExecutionState
 		return timeValue, nil
 	}
 
+	refType := ast.TypeOf(refValue)
+	if refType == ast.TMissing {
+		return time.Time{}, fmt.Errorf("reference \"%s\" is missing value", n.Node.Reference)
+	}
+
 	return time.Time{}, ErrTypeGuardFailed{RequestedType: ast.TTime, ActualType: ast.TypeOf(refValue)}
 }
 
 func (n *EvalReferenceNode) EvalDuration(scope *Scope, executionState ExecutionState) (time.Duration, error) {
-	refValue, err := n.getReferenceValue(scope, executionState)
+	refValue, err := n.getReferenceValue(scope)
 	if err != nil {
 		return 0, err
 	}
@@ -74,11 +84,16 @@ func (n *EvalReferenceNode) EvalDuration(scope *Scope, executionState ExecutionS
 		return durValue, nil
 	}
 
+	refType := ast.TypeOf(refValue)
+	if refType == ast.TMissing {
+		return 0, fmt.Errorf("reference \"%s\" is missing value", n.Node.Reference)
+	}
+
 	return 0, ErrTypeGuardFailed{RequestedType: ast.TDuration, ActualType: ast.TypeOf(refValue)}
 }
 
 func (n *EvalReferenceNode) EvalString(scope *Scope, executionState ExecutionState) (string, error) {
-	refValue, err := n.getReferenceValue(scope, executionState)
+	refValue, err := n.getReferenceValue(scope)
 	if err != nil {
 		return "", err
 	}
@@ -87,11 +102,16 @@ func (n *EvalReferenceNode) EvalString(scope *Scope, executionState ExecutionSta
 		return stringValue, nil
 	}
 
+	refType := ast.TypeOf(refValue)
+	if refType == ast.TMissing {
+		return "", fmt.Errorf("reference \"%s\" is missing value", n.Node.Reference)
+	}
+
 	return "", ErrTypeGuardFailed{RequestedType: ast.TString, ActualType: ast.TypeOf(refValue)}
 }
 
 func (n *EvalReferenceNode) EvalFloat(scope *Scope, executionState ExecutionState) (float64, error) {
-	refValue, err := n.getReferenceValue(scope, executionState)
+	refValue, err := n.getReferenceValue(scope)
 	if err != nil {
 		return float64(0), err
 	}
@@ -100,11 +120,16 @@ func (n *EvalReferenceNode) EvalFloat(scope *Scope, executionState ExecutionStat
 		return float64Value, nil
 	}
 
+	refType := ast.TypeOf(refValue)
+	if refType == ast.TMissing {
+		return float64(0), fmt.Errorf("reference \"%s\" is missing value", n.Node.Reference)
+	}
+
 	return float64(0), ErrTypeGuardFailed{RequestedType: ast.TFloat, ActualType: ast.TypeOf(refValue)}
 }
 
 func (n *EvalReferenceNode) EvalInt(scope *Scope, executionState ExecutionState) (int64, error) {
-	refValue, err := n.getReferenceValue(scope, executionState)
+	refValue, err := n.getReferenceValue(scope)
 	if err != nil {
 		return int64(0), err
 	}
@@ -113,11 +138,16 @@ func (n *EvalReferenceNode) EvalInt(scope *Scope, executionState ExecutionState)
 		return int64Value, nil
 	}
 
+	refType := ast.TypeOf(refValue)
+	if refType == ast.TMissing {
+		return int64(0), fmt.Errorf("reference \"%s\" is missing value", n.Node.Reference)
+	}
+
 	return int64(0), ErrTypeGuardFailed{RequestedType: ast.TInt, ActualType: ast.TypeOf(refValue)}
 }
 
 func (n *EvalReferenceNode) EvalBool(scope *Scope, executionState ExecutionState) (bool, error) {
-	refValue, err := n.getReferenceValue(scope, executionState)
+	refValue, err := n.getReferenceValue(scope)
 	if err != nil {
 		return false, err
 	}
@@ -126,5 +156,24 @@ func (n *EvalReferenceNode) EvalBool(scope *Scope, executionState ExecutionState
 		return boolValue, nil
 	}
 
+	refType := ast.TypeOf(refValue)
+	if refType == ast.TMissing {
+		return false, fmt.Errorf("reference \"%s\" is missing value", n.Node.Reference)
+	}
+
 	return false, ErrTypeGuardFailed{RequestedType: ast.TBool, ActualType: ast.TypeOf(refValue)}
+}
+
+func (n *EvalReferenceNode) EvalMissing(scope *Scope, executionState ExecutionState) (*ast.Missing, error) {
+	refValue, err := n.getReferenceValue(scope)
+	if err != nil {
+		return nil, err
+	}
+
+	if missingVal, isMissing := refValue.(*ast.Missing); isMissing {
+		//  This error gets checked in the eval method of a function node
+		return missingVal, fmt.Errorf("missing value: \"%v\"", n.Node.Reference)
+	}
+
+	return nil, ErrTypeGuardFailed{RequestedType: ast.TMissing, ActualType: ast.TypeOf(refValue)}
 }