+++ tick/stateful/eval_string_node.go
@@ -11,7 +11,7 @@ type EvalStringNode struct {
 	Node *ast.StringNode
 }
 
-func (n *EvalStringNode) Type(scope ReadOnlyScope, executionState ExecutionState) (ast.ValueType, error) {
+func (n *EvalStringNode) Type(scope ReadOnlyScope) (ast.ValueType, error) {
 	return ast.TString, nil
 }
 
@@ -34,12 +34,19 @@ func (n *EvalStringNode) EvalBool(scope *Scope, executionState ExecutionState) (
 func (n *EvalStringNode) EvalRegex(scope *Scope, executionState ExecutionState) (*regexp.Regexp, error) {
 	return nil, ErrTypeGuardFailed{RequestedType: ast.TRegex, ActualType: ast.TString}
 }
+
 func (n *EvalStringNode) EvalTime(scope *Scope, executionState ExecutionState) (time.Time, error) {
 	return time.Time{}, ErrTypeGuardFailed{RequestedType: ast.TTime, ActualType: ast.TString}
 }
+
 func (n *EvalStringNode) EvalDuration(scope *Scope, executionState ExecutionState) (time.Duration, error) {
 	return 0, ErrTypeGuardFailed{RequestedType: ast.TDuration, ActualType: ast.TString}
 }
+
+func (n *EvalStringNode) EvalMissing(scope *Scope, executionState ExecutionState) (*ast.Missing, error) {
+	return nil, ErrTypeGuardFailed{RequestedType: ast.TMissing, ActualType: ast.TString}
+}
+
 func (n *EvalStringNode) IsDynamic() bool {
 	return false
 }