+++ tick/stateful/scope.go
@@ -3,9 +3,26 @@ package stateful
 import (
 	"fmt"
 	"strings"
+
+	"github.com/influxdata/kapacitor/tick/ast"
 )
 
 type DynamicMethod func(self interface{}, args ...interface{}) (interface{}, error)
+
+type DynamicFunc struct {
+	F   func(args ...interface{}) (interface{}, error)
+	Sig map[Domain]ast.ValueType
+}
+
+func (df DynamicFunc) Call(args ...interface{}) (interface{}, error) {
+	return df.F(args...)
+}
+func (df DynamicFunc) Reset() {
+}
+
+func (df DynamicFunc) Signature() map[Domain]ast.ValueType {
+	return df.Sig
+}
 
 // Special marker that a value is empty
 var empty = new(interface{})
@@ -15,6 +32,7 @@ type Scope struct {
 	variables map[string]interface{}
 
 	dynamicMethods map[string]DynamicMethod
+	dynamicFuncs   map[string]*DynamicFunc
 }
 
 //Initialize a new Scope object.
@@ -22,6 +40,7 @@ func NewScope() *Scope {
 	return &Scope{
 		variables:      make(map[string]interface{}),
 		dynamicMethods: make(map[string]DynamicMethod),
+		dynamicFuncs:   make(map[string]*DynamicFunc),
 	}
 }
 
@@ -62,4 +81,12 @@ func (s *Scope) SetDynamicMethod(name string, m DynamicMethod) {
 
 func (s *Scope) DynamicMethod(name string) DynamicMethod {
 	return s.dynamicMethods[name]
+}
+
+func (s *Scope) SetDynamicFunc(name string, f *DynamicFunc) {
+	s.dynamicFuncs[name] = f
+}
+
+func (s *Scope) DynamicFunc(name string) *DynamicFunc {
+	return s.dynamicFuncs[name]
 }